+++ src/lib.rs
@@ -5,6 +5,22 @@ mod parser;
 mod tests;
 mod value;
 
-pub use parser::{load, load_dict, load_list, load_prim, ParserError, Stream};
+pub use parser::{
+    load,
+    load_dict,
+    load_list,
+    load_prim,
+    ParserError,
+    Stream,
+};
 
-pub use value::{SelectError, TraverseAction, TraverseError, Value, ValueAccessor, ValueDisplay};
+pub use value::{
+    SelectError,
+    TraverseAction,
+    TraverseError,
+    UpdateError,
+    Value,
+    ValueAccessor,
+    ValueDisplay,
+};
+