+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.1]
+
+### Changed
+
+- Implement `Display` for `UpdateError`
+- Expose `UpdateError` in the crate root
+
 ## [0.9.0] - 2020-12-31
 
 ### Added
@@ -106,7 +113,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ref variants from Value
 - Debug prints
 
-[Unreleased]: https://github.com/manokara/bencode-rs/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/manokara/bencode-rs/compare/v0.9.1...HEAD
+[0.9.1]: https://github.com/manokara/bencode-rs/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/manokara/bencode-rs/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/manokara/bencode-rs/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/manokara/bencode-rs/compare/v0.7.0...v0.8.0